+++ packages/atti-components-2/src/index.native.js
@@ -2,12 +2,12 @@ import Button from './components/Button'
 import Text from './components/Text/index.native.js'
 import ThemeProvider from './components/ThemeProvider'
 
-import theme from './theme'
+import { defaultTheme } from './themes'
 
 module.exports = {
   Button,
   Text,
   ThemeProvider,
 
-  theme,
+  defaultTheme,
 }