+++ packages/atti-components-2/src/themes/defaultTheme/componentsFactory.native.js
@@ -1,4 +1,4 @@
-import button from '../components/Button/theme'
+import button from '../../components/Button/theme'
 
 export default constants => ({
   button: button(constants),